- changed messages.ERROR's tag from error to danger - added backend validation of submission phase - added backend validation of tournament phase - added backend validation of creating test matches in tournamnent using phase - added phase validation for registration - added registration validation for submission and match creating - updated match viewing permission to incorporate all visibilities - updated match judge request validation to incorporate all visibilities - updated match and submissin deletion validation
A
- added backend validation of mathcreation p

+++ match/views.py
@@ -82,12 +82,15 @@ def judge_match(request, match_uuid):
     p = Process(target=execute_match, args=(match,))
     p.start()
 
-
     return redirectToCurrent(request)
 
 
 def delete_match(request, match_uuid):
     match = get_match_or_validate_judge_requests(request, match_uuid)
+    if match.match_visibility != Match.MatchVisibility.WORKSPACE_TEST_MATCH:
+        print("User ", request.user, " is trying to delete match ", match)
+        messages.error(request, "You can not delete this match")
+        return redirectToCurrent(request)
     try:
         os.remove(match.history_filepath)
     except IsADirectoryError:
